Don't do a density clustering.
This now just builds all of the hits into a single large cluster so that it
can be fed into TClusterSlice.  There is an ifdef to go back to the old
density cluster.  This should be split into two classes (one that just
prepares the hits into a cluster, and one the does a TDensity Cluster).
The density cluster algorithm isn't used since it is N^2 and is to slow on
reasonably sized events.