Increase error on shower clusters.
The shower clusters are not from single time slices, and include a lot of
hits coming from wires over a wide range of positions.  This means that a
simple wire weighted RMS of the positions will underestimate the position
uncertainty.  The cluster error for showers has been changed to be the
moments of the cluster charge distribution.  This probably overestimates
the uncertainty, but is more realistic than the previous estimates which
grossly underestimated the uncertainty.