Fix maximum weight to 1.0.
This changes the normalization of the weights so that the maximum weight
for any link is one, while the sum of weight*physicsWeight for a link is
also one.  The effective sets the maximum weighting for a particular hit in
a group to be physicsWeight.  For example, a hit that has a physicsWeight
of 0.5 for a particular group can contribute at most 50% of it's charge to
that group.