Big cleanup of TCluster3D.
This contiues the refactoring of TCluster3D to be a lot cleaner, and moves
several "hard coded" parameters into the parameter file.  It also
implements a new method of determining time for the combined hit based on
the overlapping charge distributions.  The hits are now split along the
time axis so that none of the 3D hits cover too much space in Z.