Refactor the track, shower, and cluster creation templates.
This reorganizes the code so that there is less duplication, and also makes
the interface more uniform for tracks and showers.  In particular, the
CreateClusters template now uses a more uniform method of splitting up the
hits into clusters so that it better matches tracks and showers.  It also
has independently tuneable parameters.