Clean up and improve documentation for CreateCluster.
Remove the code to recalculate the cluster charge based on the wire charges
since that breaks the charge sharing done during TCluster3D.  This also
changes the default for the cluster position covariance to be the moments
of the cluster charge moments.  The real uncertainty probably lies between
this assumption and the charge weighted position uncertainty.  This
assumption makes a conservative estimate of our knowledge for the position.