Handle the case when there are NO pmt hits available.
The reconstruction is only using the PMTs to determine when the event
starts, so they are not critical to the behavior.  It will now use the raw
trigger time as a t0 if there aren't any pmt hits.  It also doesn't check
if the reconstructed hits will fall into the physical "Z" of the TPC, so
out of time objects are simply reconstructed as outside of the TPC.