Make sure that the spanning tree root has edges.
The root selection code was (occasionally) selecting root vertices that
didn't have any edges attached which resulted in a "null" spanning tree.
In this case, the graph didn't change between iterations so the same root
would be chosen for the next iteration and entire chunks of the event could
be missed.