Add a user defined metric matrix for TIterativeNeighbors
The TIterativeNeighbors class uses the CGAL orthogonal iterative neighbor
algorithm to quickly (O(ln n)) find points near to a position.  The
algorithm assumes that points are in an orthogonal euclidean basis, but the
algorithm is easily adapted to a non-orthogonal basis.  With this change,
the metric can be any complete basis.